Initialize breakpoints with default content string too
This simplifies the use case of:
1. Add break point
2. Add log point to same line
